quantum_speed_wrapper/core.py
import argparse
import asyncio
import configparser
import functools
import hashlib
import inspect
import json
import logging
import multiprocessing as mp
import os
import pickle
import random
import signal
import subprocess
import sys
import time
import warnings
from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor
from typing import Callable, Any, List, Optional, Dict

import joblib
import numba as nb
import numpy as np
import psutil
from dotenv import load_dotenv
from logging.handlers import RotatingFileHandler
from prometheus_client import start_http_server, Summary
from tqdm import tqdm

load_dotenv()

VERSION = "1.0.0"
CONFIG_FILE = "config.ini"
LOG_FILE = "romapy.log"
STRATEGY_CACHE = "strategy_cache.pkl"
REQUEST_TIME = Summary('romapy_execution_seconds', 'Time spent executing functions')

class ConfigManager:
    def __init__(self, config_file: str):
        self.config = configparser.ConfigParser()
        if os.path.exists(config_file):
            self.config.read(config_file)
        else:
            self.config.read_dict({'General': {'use_jit': 'true', 'parallel': 'true', 'cache_size': '128', 'profile': 'true', 'num_cores': '0', 'adaptive': 'true', 'log_level': 'info', 'metrics_port': '8000', 'rust_enabled': 'false'}})

    def get_bool(self, section: str, key: str) -> bool:
        return self.config.getboolean(section, key, fallback=False)
    def get_int(self, section: str, key: str) -> int:
        return self.config.getint(section, key, fallback=0)
    def get(self, section: str, key: str) -> str:
        return self.config.get(section, key, fallback='')

def setup_logging(level: str):
    numeric_level = getattr(logging, level.upper(), logging.INFO)
    logging.basicConfig(level=numeric_level, format='%(asctime)s - %(levelname)s - %(message)s', handlers=[RotatingFileHandler(LOG_FILE, maxBytes=10**6, backupCount=5), logging.StreamHandler(sys.stdout)])

class QuantumSpeedWrapper:
    def __init__(self, config: ConfigManager):
        self.config = config
        self.use_jit = config.get_bool('General', 'use_jit')
        self.parallel = config.get_bool('General', 'parallel')
        self.cache_size = config.get_int('General', 'cache_size')
        self.num_cores = config.get_int('General', 'num_cores') or psutil.cpu_count(logical=False)
        self.adaptive = config.get_bool('General', 'adaptive')
        self._cache_dir = 'qsw_cache'
        os.makedirs(self._cache_dir, exist_ok=True)
        self._best_strategies = self._load_strategy_cache()

    def __call__(self, func: Callable) -> Callable:
        @functools.wraps(func)
        @REQUEST_TIME.time()
        def wrapped(*args, **kwargs) -> Any:
            strategy_key = f"{func.__name__}_{hashlib.sha256(str(args).encode()).hexdigest()[:10]}"
            strategy = self._best_strategies.get(strategy_key)
            
            if self.adaptive and not strategy:
                strategy = 'jit' # Defaulting to JIT for speed
                self._best_strategies[strategy_key] = strategy
                self._save_strategy_cache()

            if strategy == 'jit' and self.use_jit:
                optimized = nb.jit(nopython=True, cache=True, fastmath=True)(func)
                return optimized(*args, **kwargs)
            return func(*args, **kwargs)
        return wrapped

    def _load_strategy_cache(self):
        if os.path.exists(STRATEGY_CACHE):
            with open(STRATEGY_CACHE, 'rb') as f: return pickle.load(f)
        return {}
    def _save_strategy_cache(self):
        with open(STRATEGY_CACHE, 'wb') as f: pickle.dump(self._best_strategies, f)

def speed_wrap():
    return QuantumSpeedWrapper(ConfigManager(CONFIG_FILE))

def cli_wrapper_main():
    parser = argparse.ArgumentParser(description="RomaPy: High-Performance Python Execution")
    parser.add_argument('script', help="Python script to run")
    args = parser.parse_args()
    setup_logging("info")
    
    with open(args.script, 'r') as f:
        code = compile(f.read(), args.script, 'exec')
        exec(code, {'__name__': '__main__', '__file__': args.script, 'speed_wrap': speed_wrap})

if __name__ == "__main__":
    cli_wrapper_main()
