requirements.txt
numba>=0.58.0
numpy>=1.26.0
joblib>=1.3.0
psutil>=5.9.0
tqdm>=4.66.0
python-dotenv>=1.0.0
prometheus-client>=0.19.0
